+++ module/Core/src/Action/Util/ErrorResponseBuilderTrait.php
@@ -6,7 +6,7 @@ namespace Shlinkio\Shlink\Core\Action\Util;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Shlinkio\Shlink\Core\Response\NotFoundDelegate;
+use Shlinkio\Shlink\Core\Response\NotFoundHandler;
 
 trait ErrorResponseBuilderTrait
 {
@@ -14,7 +14,7 @@ trait ErrorResponseBuilderTrait
         ServerRequestInterface $request,
         RequestHandlerInterface $handler
     ): ResponseInterface {
-        $request = $request->withAttribute(NotFoundDelegate::NOT_FOUND_TEMPLATE, 'ShlinkCore::invalid-short-code');
+        $request = $request->withAttribute(NotFoundHandler::NOT_FOUND_TEMPLATE, 'ShlinkCore::invalid-short-code');
         return $handler->handle($request);
     }
 }