+++ module/Rest/config/auth.config.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Rest;
+
+return [
+
+    'auth' => [
+        'routes_whitelist' => [
+            Action\AuthenticateAction::class,
+            Action\ShortCode\SingleStepCreateShortCodeAction::class,
+        ],
+    ],
+
+];