+++ module/Rest/src/Action/ShortCode/EditShortCodeTagsAction.php
@@ -1,19 +1,23 @@
 <?php
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Rest\Action;
+namespace Shlinkio\Shlink\Rest\Action\ShortCode;
 
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Log\LoggerInterface;
 use Shlinkio\Shlink\Core\Exception\InvalidShortCodeException;
 use Shlinkio\Shlink\Core\Service\ShortUrlServiceInterface;
+use Shlinkio\Shlink\Rest\Action\AbstractRestAction;
 use Shlinkio\Shlink\Rest\Util\RestUtils;
 use Zend\Diactoros\Response\JsonResponse;
 use Zend\I18n\Translator\TranslatorInterface;
 
-class EditShortcodeTagsAction extends AbstractRestAction
+class EditShortCodeTagsAction extends AbstractRestAction
 {
+    protected const ROUTE_PATH = '/short-codes/{shortCode}/tags';
+    protected const ROUTE_ALLOWED_METHODS = [self::METHOD_PUT];
+
     /**
      * @var ShortUrlServiceInterface
      */